split ilab model list click cmds from functionality
Refactored click command handling into a dedicated `cli`
module and kept the core list functionality in a separate
module. This restructing lays the groundwork for future API
handlers and multi-processing work.
