+++ pyproject.toml
@@ -60,7 +60,7 @@ issues = "https://github.com/instructlab/instructlab/issues"
 "serve" = "instructlab.cli.model.serve:serve"
 "test" = "instructlab.model.test:test"
 "train" = "instructlab.cli.model.train:train"
-"list" = "instructlab.model.list:model_list"
+"list" = "instructlab.cli.model.list:model_list"
 
 [project.entry-points."instructlab.command.system"]
 "info" = "instructlab.system.info:info"
